telegram1/requirements.txt
# Telegram 相关依赖
telethon>=1.34.0
python-telegram-bot>=20.0

# 数据库依赖
supabase>=2.0.0

# Web API 依赖
flask>=2.3.0

# 其他依赖
httpx>=0.26.0
